refactor(app): extract file-open queue + wait tracking (phase 5g)
Move queue_file_open, process_pending_file_opens, take_completed_waits,
remove_wait_tracking, and schedule_hot_exit_recovery out of
buffer_management.rs into a new app/file_open_queue.rs sibling. ~123 lines.

CLI file arguments are queued and processed after the TUI starts so
they go through the same code path as interactive opens (proper
encoding-prompt handling). --wait tracking lets the CLI block until a
popup-based wait is dismissed.

buffer_management.rs goes from 635 -> 511 lines.
2026 tests still pass.
